Fix PHP Gherkin->Messages dependency (#1943)
When messages 18.0.0 was released, the constraint in gherkin/php/composer.json was changed to `dev-main||^17.0,^18.0`. This is incorrect because `,` is AND. It should have been `dev-main||^17.0||^18.0`

Pre-release of gherkin 23.0.0 removed the dev-main and left it as `^17.0,^18.0`. Despite visual inspection by me this was approved and released. It is currently uninstallable.

This removes the script that got rid of 'dev-main', fixes the constraint by removing `^17.0` as there is no php tag for that, and updates the messages script that adds new versions. Fixing this will require a `23.0.1` tag :/

The other issue was builds on main failing. This is because the local 'path repository' that makes gherkin install messages from local filesystem is unable to tell what version the checkout is. It's now forced to be 18.0.0 and only needs updating when gherkin breaks backward support for messages

Co-authored-by: MergeBot <ciaran.mcnulty@panasonic.aero>